外部接続を管理するexternalManager追加
