+++ Artwork_Management/ViewModels/LogInViewModel.swift
@@ -450,38 +450,6 @@ class LogInViewModel: ObservableObject {
         return UIImage(cgImage: cgImage)
     }
     
-    func shareApp(){
-        let productURL:URL = URL(string: "https://apps.apple.com/us/app/unico/id1663765686")!
-        
-        let activityViewController = UIActivityViewController(
-            activityItems: [productURL],
-            applicationActivities: nil)
-        
-        let scenes = UIApplication.shared.connectedScenes
-        let windowScene = scenes.first as? UIWindowScene
-        let window = windowScene?.windows.first
-        
-        if let window {
-            window.rootViewController?.present(activityViewController, animated: true, completion: nil)
-        }
-    }
-    
-    func reviewApp(){
-            let productURL:URL = URL(string: "https://apps.apple.com/us/app/unico/id1663765686")!
-            
-            var components = URLComponents(url: productURL, resolvingAgainstBaseURL: false)
-            
-            components?.queryItems = [
-                URLQueryItem(name: "action", value: "write-review")
-            ]
-            
-            guard let writeReviewURL = components?.url else {
-                return
-            }
-            
-            UIApplication.shared.open(writeReviewURL)
-        }
-    
     func logOut() {
         do {
             try Auth.auth().signOut()
